+++ sdks/go/node/api/client/data_ref.go
@@ -17,7 +17,7 @@ func (c apiClient) ListDescendants(
 	[]*model.DirEntry,
 	error,
 ) {
-	path := strings.Replace(api.URLPkgs_Ref_Contents, "{ref}", url.PathEscape(req.PkgRef), 1)
+	path := strings.Replace(api.URLData_Ref, "{ref}", url.PathEscape(req.DataRef), 1)
 
 	httpResp, err := c.getWithAuth(ctx, path, req.PullCreds)
 	if err != nil {