+++ sdks/go/node/api/handler/pkgs/doc.go
@@ -1,2 +1,0 @@
-// Package pkgs exposes functionality for handling "pkgs" requests.
-package pkgs