+++ sdks/go/node/core/listDescendants.go
@@ -2,6 +2,7 @@ package core
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/opctl/opctl/sdks/go/model"
 )
@@ -13,11 +14,11 @@ func (c core) ListDescendants(
 	[]*model.DirEntry,
 	error,
 ) {
-	if req.PkgRef == "" {
-		return []*model.DirEntry{}, nil
+	if req.DataRef == "" {
+		return []*model.DirEntry{}, fmt.Errorf(`"" not a valid data ref`)
 	}
 
-	dataHandle, err := c.ResolveData(ctx, req.PkgRef, req.PullCreds)
+	dataHandle, err := c.ResolveData(ctx, req.DataRef, req.PullCreds)
 	if err != nil {
 		return nil, err
 	}